end grades

+++ db/migrate/20180207190220_create_assistances.rb
@@ -3,7 +3,7 @@ class CreateAssistances < ActiveRecord::Migration[5.1]
     create_table :assistances do |t|
       t.references :enrollment, foreign_key: true
       t.references :test, foreign_key: true
-      t.decimal :grade
+      t.float :grade
 
       t.timestamps
     end
